refactor: rename files to fit new naming convention
